+++ simulator/src/cache_controller/instruction/MemoryAccess.java
@@ -14,7 +14,7 @@ public class MemoryAccess extends Instruction {
 
     public long[] adress;
     
-    public MemoryAccess(String description, int thread, long instructionAdress) {
+    public MemoryAccess(String description, long thread, long instructionAdress) {
         super(description, thread, instructionAdress);
         
         String[] splitDescr = description.split(" ");
